delete googletranslate file , fix the issue of state was not change , layput fix in mobile view also
